pyproject.toml
[project]
name = "symbol-grounding"
version = "0.1.0"
description = "Symbol-grounded image generation pipeline scaffold."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.0",
  "numpy>=1.23",
  "Pillow>=9.0",
  "diffusers>=0.24",
  "transformers>=4.35",
  "accelerate>=0.23",
  "safetensors>=0.3",
]

[project.scripts]
sg-generate = "symbol_grounding.scripts.generate_image:main"
sg-generate-diffusion = "symbol_grounding.scripts.generate_diffusion:main"
sg-edit-inpaint = "symbol_grounding.scripts.edit_inpaint:main"
sg-make-mask = "symbol_grounding.scripts.make_mask_from_layout:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["symbol_grounding*"]
